Remove executor from compilation step
